refactor: remove tui folder; move components folder to package root
