+++ src/harlequin/components/__init__.py
@@ -1,0 +1,32 @@
+from harlequin.components.code_editor import CodeEditor, EditorCollection
+from harlequin.components.error_modal import ErrorModal
+from harlequin.components.export_screen import (
+    CSVOptions,
+    ExportOptions,
+    ExportScreen,
+    JSONOptions,
+    ParquetOptions,
+)
+from harlequin.components.help_screen import HelpScreen
+from harlequin.components.messages import CursorMoved, ScrollOne
+from harlequin.components.results_viewer import ResultsViewer
+from harlequin.components.run_query_bar import RunQueryBar
+from harlequin.components.schema_viewer import CatalogItem, SchemaViewer
+
+__all__ = [
+    "CodeEditor",
+    "EditorCollection",
+    "CursorMoved",
+    "ErrorModal",
+    "ExportOptions",
+    "CSVOptions",
+    "ParquetOptions",
+    "JSONOptions",
+    "ExportScreen",
+    "HelpScreen",
+    "ResultsViewer",
+    "RunQueryBar",
+    "CatalogItem",
+    "SchemaViewer",
+    "ScrollOne",
+]
