+++ src/harlequin/components/error_modal.py
