--- src/harlequin/tui/components/__init__.py
@@ -1,32 +1,0 @@
-from harlequin.tui.components.code_editor import CodeEditor, EditorCollection
-from harlequin.tui.components.error_modal import ErrorModal
-from harlequin.tui.components.export_screen import (
-    CSVOptions,
-    ExportOptions,
-    ExportScreen,
-    JSONOptions,
-    ParquetOptions,
-)
-from harlequin.tui.components.help_screen import HelpScreen
-from harlequin.tui.components.messages import CursorMoved, ScrollOne
-from harlequin.tui.components.results_viewer import ResultsViewer
-from harlequin.tui.components.run_query_bar import RunQueryBar
-from harlequin.tui.components.schema_viewer import CatalogItem, SchemaViewer
-
-__all__ = [
-    "CodeEditor",
-    "EditorCollection",
-    "CursorMoved",
-    "ErrorModal",
-    "ExportOptions",
-    "CSVOptions",
-    "ParquetOptions",
-    "JSONOptions",
-    "ExportScreen",
-    "HelpScreen",
-    "ResultsViewer",
-    "RunQueryBar",
-    "CatalogItem",
-    "SchemaViewer",
-    "ScrollOne",
-]
