+++ tests/functional_tests/test_app.py
@@ -2,14 +2,14 @@ from pathlib import Path
 
 import pytest
 from harlequin import Harlequin
-from harlequin.tui.components import ErrorModal, ExportScreen
-from harlequin.tui.components.results_viewer import ResultsTable
+from harlequin.components import ErrorModal, ExportScreen
+from harlequin.components.results_viewer import ResultsTable
 from textual.geometry import Offset
 
 
 @pytest.fixture(autouse=True)
 def no_use_buffer_cache(monkeypatch: pytest.MonkeyPatch) -> None:
-    monkeypatch.setattr("harlequin.tui.components.code_editor.load_cache", lambda: None)
+    monkeypatch.setattr("harlequin.components.code_editor.load_cache", lambda: None)
     monkeypatch.setattr("harlequin.app.write_cache", lambda *_: None)
